uMario

+++ game/sdl2game/TextureManager.h
@@ -7,6 +7,7 @@
 #include <map>
 #include <type_traits>
 #include <SDL.h>
+#include <iostream>
 
 using namespace std;
 
@@ -23,10 +24,10 @@ public:
 
     bool load(const string &file, const string &texture_id, SDL_Renderer *renderer);
 
-    void draw(const string &texture_id, uint8_t xpos, uint8_t ypos, uint16_t height, uint16_t width,
+    void draw(const string &texture_id, int xpos, int ypos, uint16_t height, uint16_t width,
               SDL_Renderer *renderer, SDL_RendererFlip flip = SDL_FLIP_NONE);
 
-    void draw_frame(const string &texture_id, uint8_t xpos, uint8_t ypos, uint16_t height, uint16_t width,
+    void draw_frame(const string &texture_id, int xpos, int ypos, uint16_t height, uint16_t width,
                     uint16_t curr_row, uint16_t curr_frame,
                     SDL_Renderer *renderer, uint16_t offset = 70, SDL_RendererFlip flip = SDL_FLIP_NONE);
